Corrected YSorting of all objects. Added TowerSteps frenzy.
Finally fixed the YSorting of objects, which now is Towers,Enemies,PowerUps,Traps. It is no longer based on z_index, much better.
Now TowerSteps is a member of the Towers.gd class, with the appropiate methods overriden. In particular crazy shot shots only half the arrows.
Minor fixes on tower placement bug.